refactor: 세션 페이지 타입과 테스트 코드 세션 페이지 폴더로 이동
