+++ frontend/src/pages/SessionPage/hooks/__test__/mocks/useSession.mock.ts
@@ -1,4 +1,7 @@
-import { MockPeerConnections, MockSocket } from "../../type/session.test";
+import {
+  MockPeerConnections,
+  MockSocket,
+} from "@/pages/SessionPage/type/session.test";
 
 export const mockSocket: MockSocket = {
   emit: jest.fn(),
