+++ frontend/src/pages/SessionPage/hooks/useStudy.ts
@@ -1,6 +1,6 @@
 import { STUDY_EMIT_EVENT } from "@/constants/WebSocket/StudyEvent";
 import { Socket } from "socket.io-client";
-import { RoomMetadata } from "@hooks/type/session";
+import { RoomMetadata } from "@/pages/SessionPage/type/session";
 
 const useStudy = (
   socket: Socket | null,
